🎨 style(style.css): adjust button styles and add htmx indicator The button styles have been adjusted to improve the layout and display of the contact form. The width of the .btnContact class has been changed to 9rem and display has been set to flex. The .btnContactSubmit class has been removed and its styles have been merged with the .btnContact class. An htmx indicator has been added to the end of the file with a width and height of 40px.
🎨 style(public): add loader SVG
The commit adds a new SVG file for a loader animation. The SVG is generated by https://loading.io/ and is used to indicate loading state in the application.

🐛 fix(HomeController.php): add return statement to handle form submission
✨ feat(HomeController.php): add artist suggestions to the response
The HomeController.php file was missing a return statement to handle the form submission. The artist suggestions are now added to the response by rendering the artists.html.twig template. The ArtistFormType.php file now has an updated submit button with hx attributes to handle the form submission via AJAX.

🆕 feat(base.html.twig): add htmx.js library
🆕 feat(artists.html.twig): add artist suggestions template
The htmx.js library has been added to the base template to enable dynamic content loading. The artist suggestions template has also been added to the home directory. This template displays a list of suggested artists if the artistSuggestions variable is not null.

🎨 style(index.html.twig): add loading indicator to artist search button
💄 style(index.html.twig): remove artist suggestions list
The artist search button now has a loading indicator to provide feedback to the user that the search is in progress. The artist suggestions list has been removed as it was not being used and was causing confusion for users.