🔒 chore(security.yaml): comment out unused users_in_memory provider 🔒 chore(services.yaml): add PASSWORD_PROMPT parameter The users_in_memory provider is not used in the application, so it has been commented out to improve readability. A new PASSWORD_PROMPT parameter has been added to the services.yaml configuration file to allow for the configuration of a custom password prompt.
🔒 feat(HomeController.php): add password prompt feature
The password prompt feature has been added to the HomeController. If the user is not authenticated, they will be redirected to the password prompt page. The user will be prompted to enter a password, which will be compared to the password stored in the environment variables. If the passwords match, the user will be authenticated and redirected to the home page. If the passwords do not match, an error message will be displayed.

✨ feat(password_prompt.html.twig): add password prompt template
This commit adds a new template for a password prompt page. The template extends the base.html.twig template and includes a title block and a body block. The body block contains a form with a password input field and a submit button. The template also includes a for loop that displays any error messages that are passed to it via the app.flashes('error') function.